gumbo: Fix MSVC warnings (64 bit)

+++ src/gumbo/tokenizer.c
@@ -377,7 +377,7 @@ static bool temporary_buffer_equals(GumboParser* parser, const char* text) {
   // TODO(jdtang): See if the extra strlen is a performance problem, and replace
   // it with an explicit sizeof(literal) if necessary.  I don't think it will
   // be, as this is only used in a couple of rare states.
-  int text_len = strlen(text);
+  size_t text_len = strlen(text);
   return text_len == buffer->length &&
          memcmp(buffer->data, text, text_len) == 0;
 }
@@ -751,7 +751,7 @@ static void finish_tag_name(GumboParser* parser) {
   GumboTagState* tag_state = &tokenizer->_tag_state;
 
   tag_state->_tag =
-      gumbo_tagn_enum(tag_state->_buffer.data, tag_state->_buffer.length);
+      gumbo_tagn_enum(tag_state->_buffer.data, (unsigned)tag_state->_buffer.length);
   reinitialize_tag_buffer(parser);
 }
 
@@ -839,7 +839,7 @@ static bool is_appropriate_end_tag(GumboParser* parser) {
   assert(!tag_state->_is_start_tag);
   return tag_state->_last_start_tag != GUMBO_TAG_LAST &&
          tag_state->_last_start_tag == gumbo_tagn_enum(tag_state->_buffer.data,
-                                           tag_state->_buffer.length);
+                                           (unsigned)tag_state->_buffer.length);
 }
 
 void gumbo_tokenizer_state_init(
